signout api & admin dashboard

+++ admin-app/src/actions/auth.actions.js
@@ -51,7 +51,19 @@ export const isUserLoggedIn = () => {
 
 export const signout = () => {
   return async (dispatch) => {
-    localStorage.clear();
     dispatch({ type: authConstants.LOGOUT_REQUEST });
+    const res = await axios.post('/admin/signout');
+    if (res.status == 200) {
+      localStorage.clear();
+      dispatch({
+        type: authConstants.LOGOUT_SUCCESS,
+        payload: { message: res.message },
+      });
+    } else {
+      dispatch({
+        type: authConstants.LOGOUT_FAILURE,
+        payload: { error: res.data.error },
+      });
+    }
   };
 };
